Decouple polyscope raygen+cameras & move functionality to kaolin (#43)
This MR is a preparation for the eventual headless mode of the playground. The motivation is to allow additional viewers where polyscope cannot be installed, as well as easier integration within other projects.

`utils/kaolin_future` includes logic that will high probability be relocated to the next version of kaolin (v18.0).

1. Playground now assumes kaolin's Cameras.
2. Playground rendering is offered in 2 ways:  quick render_pass() and full high quality frame render()
3. VideoRecorder have moved from Playground class and reimplemented via a new interpolation module.